requirements.txt
# FastMCP 框架（新增）
fastmcp>=0.4.1

# 核心 MCP 框架（保留兼容性）
mcp>=1.0.0

# 模板引擎
jinja2>=3.1.0

# 环境变量管理
python-dotenv>=1.0.0

# 类型验证（FastMCP 需要）
pydantic>=2.0.0

# 测试框架
pytest>=7.0.0
pytest-asyncio>=0.21.0

# 类型检查（开发时推荐）
mypy>=1.0.0

# 代码格式化（开发时推荐）
black>=23.0.0
flake8>=6.0.0

# 性能分析（可选）
memory-profiler>=0.60.0
psutil>=5.9.0

# 日志增强（可选）
coloredlogs>=15.0.0